requirements.txt
Django==5.0.4
djangorestframework==3.15.1
openai==1.23.6
django-cors-headers==3.10.1
channels==3.0.4
Requests==2.31.0
requests_oauthlib==2.0.0
